The initial commit
I've learned a lot and I'm starting from scratch. I've renamed buildkit to buildkit-prev, and I'll be recreating the template here, with the new file structure I'm using. Starting with the mixins file, since I just added a new mixin to one of my projects and I want to use it in everything.